CMIS-792: Browser Binding: current state of write support
git-svn-id: https://svn.apache.org/repos/asf/chemistry/objectivecmis/branches/browser-binding@1590634 13f79535-47bb-0310-9956-ffa450edef68

+++ ObjectiveCMIS/Bindings/AtomPub/AtomPubParser/CMISAtomEntryWriter.m
@@ -200,7 +200,7 @@
             }
             default:
             {
-                CMISLogDebug(@"Property type did not match: %d", propertyData.type);
+                CMISLogDebug(@"Property type did not match: %d", (int)propertyData.type);
                 break;
             }
         }
