+++ ObjectiveCMIS/Utils/CMISHttpRequest.m
@@ -194,7 +194,7 @@ NSString * const kCMISExceptionVersioning              = @"versioning";
 - (BOOL)checkStatusCodeForResponse:(CMISHttpResponse *)response httpRequestMethod:(CMISHttpRequestMethod)httpRequestMethod error:(NSError **)error
 {
     if ([CMISLog sharedInstance].logLevel == CMISLogLevelTrace) {
-        CMISLogTrace(@"Response status code: %d", response.statusCode);
+        CMISLogTrace(@"Response status code: %ld", (long)response.statusCode);
         CMISLogTrace(@"Response body: %@", [[NSString alloc] initWithData:response.data encoding:NSUTF8StringEncoding]);
     }
